Create write module
