+++ swarmcg/scoring/__init__.py
@@ -3,3 +3,4 @@ from .bonds import get_AA_bonds_distrib, get_CG_bonds_distrib
 from .dihedrals import get_AA_dihedrals_distrib, get_CG_dihedrals_distrib
 from .sasa import compute_SASA
 from .rg import compute_Rg
+from .distances import create_bins_and_dist_matrices
